requirements.txt
appdirs==1.4.4
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cs50==9.2.2
cssselect==1.2.0
distlib==0.3.6
Django==4.1.7
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fake-useragent==1.1.1
filelock==3.8.0
Flask==2.2.2
Flask-Session==0.4.0
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
numpy==1.23.4
opencv-python==4.7.0.68
parse==1.19.0
pathlib==1.0.1
platformdirs==2.5.2
psycopg2==2.9.6
pycodestyle==2.10.0
pyee==8.2.2
PyJWT==2.6.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.1
requests-html==0.10.0
scipy==1.10.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
termcolor==2.1.0
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.13
virtualenv==20.16.6
w3lib==2.1.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.11.0
